chore: organize palette-types
- move palette api type definitons into protocol folder
- split palette api type definitons into seperate files
- TG-216
