+++ palette-types/src/Protocol/PaletteAPIError.ts
@@ -1,0 +1,7 @@
+/**
+ * Defines a consistent error type for use within the application.
+ */
+export interface PaletteAPIError {
+  param: string;
+  msg: string;
+}
