+++ palette-types/src/Types/Criteria.ts
