+++ palette-types/src/Types/Rating.ts
