+++ palette-types/src/Types/Rubric.ts
@@ -5,7 +5,7 @@
 import { Criteria } from "./Criteria";
 
 export interface Rubric {
-  id?: number; // id is optional as new rubrics will not have one assigned by Canvas yet
+  id?: number; // OPTIONAL: new rubrics will not have one assigned by Canvas
   title: string;
   pointsPossible: number;
   key: string; // required unique id for react
